feat: characterize level sets of the order for analytic/meromorphic functions (#22292)
Establish that sets where the order of analytic/meromorphic function takes values other than infinity and zero are codiscrete within the domains of analyticity/meromorphicity. Uniformize API between analytic and meromorphic functions. Fix typo in docstrings.

This material is used in [Project VD](https://github.com/kebekus/ProjectVD), which aims to formalize Value Distribution Theory for meromorphic functions on the complex plane. It will be used in the definition of a divisor attached to a meromorphic function.